requirements.txt
streamlit
google-generativeai
google-ai-generativelanguage